pyproject.toml
[project]
name = "pyspark-cr7"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.7"
dependencies = [
    "black>=25.9.0",
    "click>=8.3.0",
    "delta-spark>=4.0.0",
    "notebook>=7.4.7",
    "pyspark>=4.0.1",
]
